% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmse.R
\name{rmse}
\alias{rmse}
\title{Find Root Mean Square Error}
\usage{
rmse(actual, predicted)
}
\arguments{
\item{actual}{Actual observations}

\item{predicted}{Predicted Values}
}
\value{
rmse root mean squared error
}
\description{
This function finds root mean square error between
actual observations and predicted values. The actual
and predicted input vectors must have same number of
rows
}
